cyberbullying-detection-api/pyproject.toml
[tool.poetry]
name = "cyberbullying-detection-api"
version = "0.1.0"
description = ""
authors = ["rzkmak <rizkimaulanaakbar@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
scipy = "^1.12.0"
numpy = "^1.26.3"
torch = "^2.2.0"
transformers = "^4.37.2"
tensorflow-io-gcs-filesystem = "0.34.0"
tensorflow-macos = "^2.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
